@@ -1,20 +1,20 @@
-String stationIdDest;
-int stationIdSource;
-bool isInsideUSTHB;
+String? stationIdDest;
+int? stationIdSource;
+bool? isInsideUSTHB;
 int postorGet=0;
 int globalIndex=0;
 
-String velo_name='';
-int velo_password=0;
-int velo=-1;
+String? velo_name='';
+int? velo_password=0;
+int? velo=-1;
 String response='';
 bool isConnectedtoBT =false;
 bool reserved = false;
 String authString='';
-Duration total_ride_time;
+Duration? total_ride_time;
 
-int new_solde=0;
-int total_price=0;
+int? new_solde=0;
+int? total_price=0;
 
 String bt_status='';
 
