@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import '/widgets/onboarding_widget.dart';
 
 class Onboarding extends StatelessWidget {
-  const Onboarding({Key key}) : super(key: key);
+  const Onboarding({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
