@@ -4,7 +4,7 @@ import '/models/db.dart';
 import '/helpers/globals.dart' as globals;
 
 
-Future<double> refillSold(String token, String voucher) async {
+Future<double?> refillSold(String? token, String voucher) async {
   final url = Uri.parse('${globals.api}/sold/');
   final headers = {'Authorization': 'Token $token'};
   final body = {'token': voucher};
@@ -12,7 +12,7 @@ Future<double> refillSold(String token, String voucher) async {
 
   if (response.statusCode == 200) {
     final data = jsonDecode(response.body);
-    final double sold = data['sold'];
+    final double? sold = data['sold'];
     return sold;
   } else {
     throw Exception('Failed to retrieve sold: ${response.statusCode}');
